charts update

+++ winch/models.py
@@ -31,7 +31,7 @@ class WinchDataLog(models.Model):
     tetherline_length = models.FloatField(blank=True, null=True, help_text='티더선 길이')
     tetherline_angle = models.FloatField(blank=True, null=True, help_text='티더선 각도')
     tetherline_tension = models.FloatField(blank=True, null=True, help_text='티더선 장력')
-    pressure = models.IntegerField(blank=True, null=True, help_text='기압')
+    pressure = models.FloatField(blank=True, null=True, help_text='기압')
     temperature = models.FloatField(blank=True, null=True, help_text='온도')
     wind_direction = models.CharField(max_length=100, blank=True, null=True, help_text='풍향')
     wind_speed = models.FloatField(blank=True, null=True, help_text='풍속')
